[fix] rename data_structures namespace (#1061)
* rename data_structures namespace

* spelling correction

* fix namespace doc name
